Add console=ttyS0 and console=tty0 to cmdline
The AMD Milan hardware we have access to has the consoles plugged into
ttyS0. So we need to support both ttyS0 and ttyS1. Further, tty0 is the
default from the kernel's perspective. So we add it at the end to ensure
expected behavior (see the kernel docs).

The end result of this patch is that we will get input on tty0 and
output on tty0, ttyS0 and ttyS1. This should work for all the configured
systems.

Signed-off-by: Nathaniel McCallum <nathaniel@congru.us>